style: ajusta formatação com forge fmt

+++ src/VendaIngressosNFT.sol
@@ -5,39 +5,37 @@ import "openzeppelin-contracts/contracts/token/ERC721/extensions/ERC721URIStorag
 import "openzeppelin-contracts/contracts/access/Ownable.sol";
 
 contract VendaIngressosNFT is ERC721URIStorage, Ownable {
-    uint public proximoIdIngresso = 1;
+    uint256 public proximoIdIngresso = 1;
 
-    enum TipoVenda { Aberta, PorConvite }
+    enum TipoVenda {
+        Aberta,
+        PorConvite
+    }
 
     struct Evento {
         string nome;
         address organizador;
-        uint preco;
-        uint totalIngressos;
-        uint vendidos;
+        uint256 preco;
+        uint256 totalIngressos;
+        uint256 vendidos;
         TipoVenda tipo;
         mapping(address => bool) convidados;
-        mapping(address => uint) revendedores; // autorizados com quantidade
+        mapping(address => uint256) revendedores; // autorizados com quantidade
     }
 
-    uint public proximoIdEvento = 1;
-    mapping(uint => Evento) private eventos;
-    mapping(uint => uint) public ingressoParaEvento;
+    uint256 public proximoIdEvento = 1;
+    mapping(uint256 => Evento) private eventos;
+    mapping(uint256 => uint256) public ingressoParaEvento;
 
-    event EventoCriado(uint idEvento, string nome, address organizador);
-    event IngressoComprado(uint idEvento, address comprador, uint tokenId);
-    event IngressosRepasados(uint idEvento, address para, uint quantidade);
-    event IngressoRevendido(uint tokenId, address de, address para);
+    event EventoCriado(uint256 idEvento, string nome, address organizador);
+    event IngressoComprado(uint256 idEvento, address comprador, uint256 tokenId);
+    event IngressosRepasados(uint256 idEvento, address para, uint256 quantidade);
+    event IngressoRevendido(uint256 tokenId, address de, address para);
 
     constructor() ERC721("IngressoNFT", "ING") Ownable(msg.sender) {}
 
-    function criarEvento(
-        string memory _nome,
-        uint _preco,
-        uint _totalIngressos,
-        TipoVenda _tipo
-    ) external {
-        uint idEvento = proximoIdEvento;
+    function criarEvento(string memory _nome, uint256 _preco, uint256 _totalIngressos, TipoVenda _tipo) external {
+        uint256 idEvento = proximoIdEvento;
         Evento storage novo = eventos[idEvento];
         novo.nome = _nome;
         novo.organizador = msg.sender;
@@ -49,17 +47,17 @@ contract VendaIngressosNFT is ERC721URIStorage, Ownable {
         emit EventoCriado(idEvento, _nome, msg.sender);
     }
 
-    function adicionarConvidado(uint _idEvento, address _convidado) external {
+    function adicionarConvidado(uint256 _idEvento, address _convidado) external {
         Evento storage evento = eventos[_idEvento];
         require(msg.sender == evento.organizador, "Apenas o organizador pode convidar");
         evento.convidados[_convidado] = true;
     }
 
-    function estaConvidado(uint _idEvento, address _usuario) public view returns (bool) {
+    function estaConvidado(uint256 _idEvento, address _usuario) public view returns (bool) {
         return eventos[_idEvento].convidados[_usuario];
     }
 
-    function comprarIngresso(uint _idEvento, string memory tokenURI) external payable {
+    function comprarIngresso(uint256 _idEvento, string memory tokenURI) external payable {
         Evento storage evento = eventos[_idEvento];
 
         require(evento.organizador != address(0), "Evento inexistente");
@@ -70,7 +68,7 @@ contract VendaIngressosNFT is ERC721URIStorage, Ownable {
             require(evento.convidados[msg.sender], "Voce nao esta na lista de convidados");
         }
 
-        uint tokenId = proximoIdIngresso;
+        uint256 tokenId = proximoIdIngresso;
         _safeMint(msg.sender, tokenId);
         _setTokenURI(tokenId, tokenURI);
 
@@ -81,13 +79,13 @@ contract VendaIngressosNFT is ERC721URIStorage, Ownable {
         emit IngressoComprado(_idEvento, msg.sender, tokenId);
     }
 
-    function revenderIngresso(uint tokenId, address para) external {
+    function revenderIngresso(uint256 tokenId, address para) external {
         require(ownerOf(tokenId) == msg.sender, "Voce nao possui este ingresso");
         _transfer(msg.sender, para, tokenId);
         emit IngressoRevendido(tokenId, msg.sender, para);
     }
 
-    function repassarIngressos(uint _idEvento, address para, uint quantidade) external {
+    function repassarIngressos(uint256 _idEvento, address para, uint256 quantidade) external {
         Evento storage evento = eventos[_idEvento];
         require(msg.sender == evento.organizador, "Apenas o organizador pode repassar");
         require(evento.vendidos + quantidade <= evento.totalIngressos, "Nao ha ingressos suficientes");
@@ -98,17 +96,17 @@ contract VendaIngressosNFT is ERC721URIStorage, Ownable {
         emit IngressosRepasados(_idEvento, para, quantidade);
     }
 
-    function retirarRevendedor(uint _idEvento, address revendedor) external {
+    function retirarRevendedor(uint256 _idEvento, address revendedor) external {
         Evento storage evento = eventos[_idEvento];
         require(msg.sender == evento.organizador, "Apenas o organizador pode remover");
         evento.revendedores[revendedor] = 0;
     }
 
-    function venderComoRevendedor(uint _idEvento, string memory tokenURI) external {
+    function venderComoRevendedor(uint256 _idEvento, string memory tokenURI) external {
         Evento storage evento = eventos[_idEvento];
         require(evento.revendedores[msg.sender] > 0, "Sem ingressos disponiveis para revenda");
 
-        uint tokenId = proximoIdIngresso;
+        uint256 tokenId = proximoIdIngresso;
         _safeMint(msg.sender, tokenId);
         _setTokenURI(tokenId, tokenURI);
 
@@ -119,29 +117,25 @@ contract VendaIngressosNFT is ERC721URIStorage, Ownable {
         emit IngressoComprado(_idEvento, msg.sender, tokenId);
     }
 
-
-    function sacarFundos(uint _idEvento) external {
+    function sacarFundos(uint256 _idEvento) external {
         Evento storage evento = eventos[_idEvento];
         require(msg.sender == evento.organizador, "Apenas o organizador pode sacar");
         payable(evento.organizador).transfer(address(this).balance);
     }
 
-    function obterDadosEvento(uint _idEvento) external view returns (
-        string memory nome,
-        address organizador,
-        uint preco,
-        uint totalIngressos,
-        uint vendidos,
-        TipoVenda tipo
-    ) {
+    function obterDadosEvento(uint256 _idEvento)
+        external
+        view
+        returns (
+            string memory nome,
+            address organizador,
+            uint256 preco,
+            uint256 totalIngressos,
+            uint256 vendidos,
+            TipoVenda tipo
+        )
+    {
         Evento storage evento = eventos[_idEvento];
-        return (
-            evento.nome,
-            evento.organizador,
-            evento.preco,
-            evento.totalIngressos,
-            evento.vendidos,
-            evento.tipo
-        );
+        return (evento.nome, evento.organizador, evento.preco, evento.totalIngressos, evento.vendidos, evento.tipo);
     }
 }
